reorganize header files into hierarchical structure
