+++ include/raccoon/internals/polynomial/challenge.hpp
