+++ include/raccoon/internals/polynomial/masked_poly.hpp
