+++ include/raccoon/internals/polynomial/poly_mat.hpp
