+++ tests/test_gadgets.cpp
@@ -1,4 +1,4 @@
-#include "raccoon/internals/poly_vec.hpp"
+#include "raccoon/internals/polynomial/poly_vec.hpp"
 #include <gtest/gtest.h>
 
 template<size_t rows, size_t d>
